requirement.txt
pandas
numpy
faiss-cpu
sentence-transformers
transformers
torch
